backend-new/requirements.txt
asgiref==3.8.1
Django==4.2.14
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
pyJWT==2.6.0
Rust==1.3.1
pysam==0.22.1
Pillow==10.4.0
django-cleanup==8.1.0
cffi==1.16.0
cryptography==42.0.8
pycparser==2.22
psycopg2-binary==2.9.9
channels==4.1.0
requests==2.32.3
google-auth-oauthlib==1.2.1
oauthlib==3.2.2
requests-oauthlib==2.0.0
pipdeptree==2.23.1
drf-spectacular==0.27.0
amqp==5.2.0
django-unfold==0.20.1
pika
whitenoise